docs(marbles): move marble scripts over to docs app
