+++ docs_app/package.json
@@ -1,8 +1,8 @@
 {
-  "name": "rxjs.io",
+  "name": "rxjs.dev",
   "version": "1.0.0",
   "main": "index.js",
-  "repository": "git@github.com:ReactiveX/  rxjs",
+  "repository": "git@github.com:ReactiveX/rxjs",
   "author": "RxJS",
   "license": "MIT",
   "scripts": {
@@ -37,7 +37,8 @@
     "build-ie-polyfills": "webpack-cli src/ie-polyfills.js -o src/generated/ie-polyfills.min.js --mode production",
     "update-webdriver": "webdriver-manager update --standalone false --gecko false $CHROMEDRIVER_VERSION_ARG",
     "~~build": "ng build --configuration=stable --prod --source-map",
-    "~~clean-generated": "node --eval \"require('shelljs').rm('-rf', 'src/generated')\""
+    "~~clean-generated": "node --eval \"require('shelljs').rm('-rf', 'src/generated')\"",
+    "build:marbles": "ts-node -P ./tools/marbles/tsconfig.marbles.json ./tools/marbles/scripts/index.ts"
   },
   "engines": {
     "node": ">=10.9"
@@ -74,6 +75,7 @@
     "@types/jasmine": "^3.3.13",
     "@types/jasminewd2": "^2.0.3",
     "@types/node": "^12.11.1",
+    "@types/svgo": "^1.3.3",
     "archiver": "^3.0.0",
     "canonical-path": "^1.0.0",
     "chalk": "^2.1.0",
@@ -122,6 +124,9 @@
     "semver": "^6.1.1",
     "shelljs": "^0.8.3",
     "svgson": "^4.1.0",
+    "svgo": "^1.3.2",
+    "swirly": "^0.13.1",
+    "swirly-types": "^0.13.1",
     "tree-kill": "^1.1.0",
     "ts-node": "^8.2.0",
     "tslint": "^5.17.0",
