+++ docs_app/tools/marbles/tsconfig.marbles.json
@@ -1,0 +1,7 @@
+{
+    "extends": "../../tsconfig.json",
+    "compilerOptions": {
+      "module": "commonjs"
+    }
+  }
+
