+++ package.json
@@ -42,8 +42,7 @@
     "compile": "tsc -b ./src/tsconfig.cjs.json ./src/tsconfig.cjs.spec.json ./src/tsconfig.esm.json ./src/tsconfig.esm5.json ./src/tsconfig.esm5.rollup.json ./src/tsconfig.types.json ./spec/tsconfig.json",
     "build:clean": "shx rm -rf ./dist",
     "build:global": "node ./tools/make-umd-bundle.js && node ./tools/make-closure-core.js",
-    "build:package": "npm-run-all build:clean compile build:global && node ./tools/prepare-package.js && node ./tools/generate-alias.js",
-    "build:marbles": "ts-node ./marbles/scripts/index.ts"
+    "build:package": "npm-run-all build:clean compile build:global && node ./tools/prepare-package.js && node ./tools/generate-alias.js"
   },
   "repository": {
     "type": "git",
@@ -107,7 +106,6 @@
     "@types/sinon": "4.1.3",
     "@types/sinon-chai": "2.7.29",
     "@types/source-map": "^0.5.2",
-    "@types/svgo": "^1.3.3",
     "babel-polyfill": "6.26.0",
     "chai": "4.1.2",
     "check-side-effects": "0.0.20",
@@ -140,11 +138,7 @@
     "sinon": "4.3.0",
     "sinon-chai": "2.14.0",
     "source-map-support": "0.5.3",
-    "svgo": "^1.3.2",
-    "swirly": "^0.13.1",
-    "swirly-types": "^0.13.1",
     "systemjs": "^0.21.0",
-    "ts-node": "^8.10.1",
     "tslint": "5.9.1",
     "tslint-etc": "1.9.2",
     "tslint-no-toplevel-property-access": "0.0.2",
